backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1

# Code analysis
gitpython==3.1.41
tree-sitter==0.20.4
radon==6.0.1

# AI
google-generativeai==0.3.2

# PDF generation
reportlab==4.0.9
PyPDF2==3.0.1

# Utilities
python-magic==0.4.27
chardet==5.2.0